<PY|TS|SU> Added test for Region Recommendation card.
