--- packages/client/src/utils/hooks/EmissionServiceHook.ts
@@ -1,26 +1,0 @@
-/* eslint-disable @typescript-eslint/no-unused-vars */
-/*
- * © 2021 Thoughtworks, Inc.
- */
-
-import { useEffect } from 'react'
-import React from 'react'
-
-const useRemoteEmissionService = (params: any): any => {
-  const [result, setResult] = React.useState<any>()
-  useEffect(() => {
-    const fetchBestLocation = async () => {
-      const response = await Promise.all(params)
-      setResult(
-        response.map((res) => {
-          return res.data
-        }),
-      )
-    }
-    fetchBestLocation()
-  }, [])
-
-  return result
-}
-
-export default useRemoteEmissionService
